All classes shall be serialized. Serialized attribute is applied only on members
Unfortunately, attributes can be automatically inherited only on methods,
but not on classes, meaning that a class inheriting from a serialized
class will not be serialized. This may be particurarly troublesome in some
cases.

Having to apply to Serialized attribute both to the class and its members felt
unnatural anway.

Keeping it simple is much better.

Note that if class types have changed after been serialized so as to
parents are now different from what use to be, there would be issues.

This is not handled by Albite.Serialization anyway.

However, if the version of the assembly differs, creating the serializer
will fail if the version of the serialized type is newer, because
deserializing the type through Type.GetType() would fail.